requirements.txt
pyTelegramBotAPI
pycups
fpdf2
Pillow